SawyerEliassenSolver

The main module of this package. It reexports all the public names defined in its submodules.

SawyerEliassenSolver.SawyerEliassenSolver — Module

SawyerEliassenSolver.jl – Solve the Sawyer-Eliassen equation using a pseudo-spectral discretisation and 3rd order accurate implicit timestepping for arbitrary balanced background flows.

Exports

AbstractForcingFunction

AbstractPreconditioner

BackgroundFlow

Clock

DiagonalQuadraticPreconditioner

Domain

Domains

FCVariable

FSVariable

FZVariable

Forcing

GlobalPhysicalForcing

GlobalSpectralForcing

Grid

IdentityPreconditioner

NoForcing

Output

OutputVariable

OutputVariables

OutputWriter

PhysicalForcing

PointwisePhysicalForcing

PointwiseSpectralForcing

Problem

Problems

Scratch

SpectralForcing

State

Timestepper

Timesteppers

Variables

XCVariable

XSVariable

XZVariable

add_output_variables!

advance!

compute!

compute_ζₜ!

consistent_domains

cosine_transform

cosine_transform!

get_domain

get_iteration

get_time

gridpoints

horizontal_counterpart

horizontal_transform

horizontal_transform!

resolved

resolved_cosine_indices

resolved_fourier_indices

resolved_sine_indices

set_b!

set_v!

set_vb!

set_ζ!

sine_transform

sine_transform!

transform!

wavenumbers

wavenumbers_full

write!

write_attributes!

xgridpoints

xwavenumbers

xwavenumbers_full

zgridpoints

zwavenumbers

zwavenumbers_full

∂x

∂x!

∂z

∂z!

∂z²

∂z²!

∇⁻²!

∫dx

∫dx!

∫dz

∫dz!

∫dz²

∫dz²!