Forcing

SawyerEliassenSolver.Forcing — Module

Module for implementing different RHS forcings.

Exports

AbstractForcing

GlobalPhysicalForcing

GlobalSpectralForcing

NoForcing

PhysicalForcing

PointwisePhysicalForcing

PointwiseSpectralForcing

SpectralForcing

SawyerEliassenSolver.Forcing.PhysicalForcing — Type

Union of types representing forcing in physical (XZ) space.

SawyerEliassenSolver.Forcing.SpectralForcing — Type

Union of types representing forcing in spectral (FS) space.

SawyerEliassenSolver.Forcing.AbstractForcing — Type

Abstract super type for all forcings.

SawyerEliassenSolver.Forcing.GlobalPhysicalForcing — Type

struct GlobalPhysicalForcing{T, P} <: AbstractForcing{T}

Apply global forcing to the Sawyer-Eliassen equation in physical space. func is a function that updates the entire forcing variable in-place at time t. func should have signature func(F::XZVariable{T},t::T,params::P)::Nothing where {T<:Union{Float32,Float64}}. unless params is nothing, in which case func should have signature func(F::XZVariable{T},t::T)::T where {T<:Union{Float32,Float64}}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.GlobalPhysicalForcing — Method

struct GlobalPhysicalForcing{T, P} <: AbstractForcing{T}

Global physical forcing with no parameters. func should have signature func(F::XZVariable{T},t::T)::T where {T<:Union{Float32,Float64}}.

SawyerEliassenSolver.Forcing.GlobalSpectralForcing — Type

struct GlobalSpectralForcing{T, P} <: AbstractForcing{T}

Apply global forcing to the Sawyer-Eliassen equation in spectral space. func is a function that updates the entire forcing variable in-place at time t. func should have signature func(F::FSVariable{T},t::T,params::P)::Nothing where {T<:Union{Float32,Float64}} unless params is nothing, in which case func should have signature func(F::FSVariable{T},t::T)::T where {T<:Union{Float32,Float64}} Note that the FSVariable{T} has entries of type Complex{T}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.GlobalSpectralForcing — Method

struct GlobalSpectralForcing{T, P} <: AbstractForcing{T}

Global spectral forcing with no parameters. func should have signature func(F::FSVariable{T},t::T)::T where {T<:Union{Float32,Float64}}.

SawyerEliassenSolver.Forcing.NoForcing — Type

Type representing no forcing.

SawyerEliassenSolver.Forcing.PointwisePhysicalForcing — Type

struct PointwisePhysicalForcing{T, P} <: AbstractForcing{T}

Apply pointwise forcing to the Sawyer-Eliassen equation in physical space. func is a function that computes the forcing at a given position and time. func should have signature func(x::T,z::T,t::T,params::P)::T where {T<:Union{Float32,Float64}} unless params is nothing, in which case func should have signature func(x::T,z::T,t::T)::T where {T<:Union{Float32,Float64}}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.PointwisePhysicalForcing — Method

struct PointwisePhysicalForcing{T, P} <: AbstractForcing{T}

Pointwise physical forcing with no parameters. func should have signature func(x::T,z::T,t::T)::T where {T<:Union{Float32,Float64}}.

SawyerEliassenSolver.Forcing.PointwiseSpectralForcing — Type

struct PointwiseSpectralForcing{T, P} <: AbstractForcing{T}

Apply pointwise forcing to the Sawyer-Eliassen equation in spectral space. func is a function that computes the forcing at given wavenumbers and time. func should have signature func(kx::T,kz::T,t::T,params::P)::complex{T} where {T<:Union{Float32,Float64}}. unless params is nothing, in which case func should have signature func(kx::T,kz::T,t::T)::T where {T<:Union{Float32,Float64}}

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.PointwiseSpectralForcing — Method

struct PointwiseSpectralForcing{T, P} <: AbstractForcing{T}

Pointwise spectral forcing with no parameters. func should have signature func(kx::T,kz::T,t::T)::T where {T<:Union{Float32,Float64}}.