Variables

SawyerEliassenSolver.Variables — Module

Module for different variable representations, transforms between representations and operators that act on variables.

Exports

FCVariable

FSVariable

FZVariable

XCVariable

XSVariable

XZVariable

cosine_transform

cosine_transform!

horizontal_counterpart

horizontal_transform

horizontal_transform!

resolved

sine_transform

sine_transform!

transform!

∂x

∂x!

∂z

∂z!

∂z²

∂z²!

∇⁻²!

∫dx

∫dx!

∫dz

∫dz!

∫dz²

∫dz²!

SawyerEliassenSolver.Variables.FCVariable — Type

A variable expressed in Fourier space in the horizontal and cosine space in the vertical.

SawyerEliassenSolver.Variables.FSVariable — Type

A variable expressed in Fourier space in the horizontal and sine space in the vertical.

SawyerEliassenSolver.Variables.FZVariable — Type

A variable expressed in Fourier space in the horizontal and physical space in the vertical.

SawyerEliassenSolver.Variables.XCVariable — Type

A variable expressed in physical space in the horizontal and cosine space in the vertical.

SawyerEliassenSolver.Variables.XSVariable — Type

A variable expressed in physical space in the horizontal and sine space in the vertical.

SawyerEliassenSolver.Variables.XZVariable — Type

A variable expressed in physical space in the horizontal and vertical.

SawyerEliassenSolver.Variables.cosine_transform — Function

Transform the variable from physical space to cosine space or vice verse in the vertical.

SawyerEliassenSolver.Variables.cosine_transform! — Method

cosine_transform!(out::XCVariable, in::XZVariable)

SawyerEliassenSolver.Variables.cosine_transform! — Method

cosine_transform!(out::XZVariable, in::XCVariable)

SawyerEliassenSolver.Variables.horizontal_counterpart — Method

Return the fourier (physical) counterpart to a physical (fourier) variable.

SawyerEliassenSolver.Variables.horizontal_transform — Function

Transform the variable from physical space to Fourier space or vice verse in the horizontal.

SawyerEliassenSolver.Variables.horizontal_transform! — Method

horizontal_transform!(
    out::Union{FCVariable{T}, FSVariable{T}, FZVariable{T}} where T,
    in::Union{XCVariable{T}, XSVariable{T}, XZVariable{T}} where T
)

SawyerEliassenSolver.Variables.horizontal_transform! — Method

horizontal_transform!(
    out::Union{XCVariable{T}, XSVariable{T}, XZVariable{T}} where T,
    in::Union{FCVariable{T}, FSVariable{T}, FZVariable{T}} where T
)

SawyerEliassenSolver.Variables.resolved — Function

Return a view into the resolved wavenumbers.

SawyerEliassenSolver.Variables.sine_transform — Function

Transform the variable from physical space to sine space or vice verse in the vertical.

SawyerEliassenSolver.Variables.sine_transform! — Method

sine_transform!(out::XSVariable, in::XZVariable)

SawyerEliassenSolver.Variables.sine_transform! — Method

sine_transform!(out::XZVariable, in::XSVariable)

SawyerEliassenSolver.Variables.transform! — Method

Safe transforms with domain validation for public use.

SawyerEliassenSolver.Variables.∂x — Function

Document me!

SawyerEliassenSolver.Variables.∂x! — Function

Document me!

SawyerEliassenSolver.Variables.∂z — Function

Document me!

SawyerEliassenSolver.Variables.∂z! — Function

Document me!

SawyerEliassenSolver.Variables.∂z² — Function

Document me!

SawyerEliassenSolver.Variables.∂z²! — Function

Document me!

SawyerEliassenSolver.Variables.∇⁻²! — Function

Document me!

SawyerEliassenSolver.Variables.∫dx — Function

Document me!

SawyerEliassenSolver.Variables.∫dx! — Function

Document me!

SawyerEliassenSolver.Variables.∫dz — Function

Document me!

SawyerEliassenSolver.Variables.∫dz! — Function

Document me!

SawyerEliassenSolver.Variables.∫dz² — Function

Document me!

SawyerEliassenSolver.Variables.∫dz²! — Function

Document me!