Forcing

SawyerEliassenSolver.Forcing.PhysicalForcing — Type

Union of types representing forcing in physical (XZ) space.

SawyerEliassenSolver.Forcing.SpectralForcing — Type

Union of types representing forcing in spectral (FS) space.

SawyerEliassenSolver.Forcing.AbstractForcingFunction — Type

Abstract super type for all forcings.

SawyerEliassenSolver.Forcing.GlobalPhysicalForcing — Type

struct GlobalPhysicalForcing{T, P} <: AbstractForcingFunction{T}

Apply global forcing to the Sawyer-Eliassen equation in physical space. func is a function that updates the entire forcing variable in-place at time t. func should have signature func(F::XZVariable{T},t::T,params::P)::Nothing where {T<:Union{Float32,Float64}}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.GlobalSpectralForcing — Type

struct GlobalSpectralForcing{T, P} <: AbstractForcingFunction{T}

Apply global forcing to the Sawyer-Eliassen equation in spectral space. func is a function that updates the entire forcing variable in-place at time t. func should have signature func(F::FSVariable{T},t::T,params::P)::Nothing where {T<:Union{Float32,Float64}}. Note that the FSVariable{T} has entries of type Complex{T}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.NoForcing — Type

Type representing no forcing.

SawyerEliassenSolver.Forcing.PointwisePhysicalForcing — Type

struct PointwisePhysicalForcing{T, P} <: AbstractForcingFunction{T}

Apply pointwise forcing to the Sawyer-Eliassen equation in physical space. func is a function that computes the forcing at a given position and time. func should have signature func(x::T,z::T,t::T,params::P)::T where {T<:Union{Float32,Float64}}.

Fields

domain::Domain{T} where T

func::Function

params::Any

SawyerEliassenSolver.Forcing.PointwiseSpectralForcing — Type

struct PointwiseSpectralForcing{T, P} <: AbstractForcingFunction{T}

Apply pointwise forcing to the Sawyer-Eliassen equation in spectral space. func is a function that computes the forcing at given wavenumbers and time. func should have signature func(kx::T,kz::T,t::T,params::P)::complex{T} where {T<:Union{Float32,Float64}}.

Fields

domain::Domain{T} where T

func::Function

params::Any